Refactor air-gapped charts for better re-use in other charts
