+++ src/main/groovy/com/cloudogu/gitops/config/ApplicationConfigurator.groovy
@@ -83,6 +83,7 @@ class ApplicationConfigurator {
                     url     : '',
                     username: DEFAULT_ADMIN_USER,
                     password: DEFAULT_ADMIN_PW,
+                    gitOpsUsername : '', // Set dynamically
                     /* This corresponds to the "Base URL" in SCMM Settings.
                        We use the K8s service as default name here, to make the build on push feature (webhooks from SCMM to Jenkins that trigger builds) work in k3d.
                        The webhook contains repository URLs that start with the "Base URL" Setting of SCMM.
@@ -106,6 +107,8 @@ class ApplicationConfigurator {
             application: [
                     remote        : false,
                     airGapped     : false,
+                    // Take from env because the Dockerfile provides a local copy of the repo for air-gapped mode
+                    localHelmChartFolder: System.getenv('LOCAL_HELM_CHART_FOLDER'),
                     insecure      : false,
                     username      : DEFAULT_ADMIN_USER,
                     password      : DEFAULT_ADMIN_PW,
@@ -180,8 +183,6 @@ class ApplicationConfigurator {
                                     repoURL: 'https://prometheus-community.github.io/helm-charts',
                                     /* When updating this make sure to also test if air-gapped mode still works */
                                     version: '58.2.1',
-                                    // Take from env because the Dockerfile provides a local copy of the repo for air-gapped mode
-                                    localFolder: System.getenv('KUBE_PROM_STACK_HELMCHART_PATH'), 
                                     grafanaImage: '',
                                     grafanaSidecarImage: '',
                                     prometheusImage: '',
@@ -351,6 +352,8 @@ class ApplicationConfigurator {
     private void addScmmConfig(Map newConfig) {
         log.debug("Adding additional config for SCM-Manager")
 
+        newConfig.scmm['gitOpsUsername'] = "${this.config.application['namePrefix']}gitops"
+
         if (newConfig.scmm["url"]) {
             log.debug("Setting external scmm config")
             newConfig.scmm["internal"] = false
@@ -464,14 +467,11 @@ class ApplicationConfigurator {
                 !configToSet.scmm["url"] && configToSet.jenkins["url"]) {
             throw new RuntimeException('When setting jenkins URL, scmm URL must also be set and the other way round')
         }
-        def monitoringHelmConfig = configToSet['features']['monitoring']['helm']
-        if (!monitoringHelmConfig['localFolder']) {
+        if (!configToSet.application['localHelmChartFolder']) {
             // This should only happen when run outside the image, i.e. during development
-            throw new RuntimeException("Missing config for localFolder of helm chart ${monitoringHelmConfig['chart']}.\n" +
+            throw new RuntimeException("Missing config for localHelmChartFolder.\n" +
                     "Either run inside the official container image or setting env var " +
-                    "KUBE_PROM_STACK_HELMCHART_PATH='charts/kube-prometheus-stack' after running this:\n" +
-                    "helm repo add prometheus-community ${monitoringHelmConfig['repoURL']}\n" +
-                    "helm pull --untar --untardir charts prometheus-community/${monitoringHelmConfig['chart']} --version ${monitoringHelmConfig['version']}")
+                    "LOCAL_HELM_CHART_FOLDER='charts' after running 'scripts/downloadHelmCharts.sh' from the repo")
         }
 
     }
